Update(10/5/17)Health and Healing DONE!
- new getmaxHitpoint() method is introduced to the SWEntityInterface.
This means that for every entity we assume that the initialized
hitpoints is the maximum hitpoint they could have.

- accounted for when Ben's hitpoint is less than the canteen's hitpoint:
assuming that ben's hitpoint is now 96, but for every drink from the
canteen, it restores 10 hitpoints, thus when ben's hitpoints is 96 we
should only restore 4 hitpoints. 

- HEALTH and HEALING DONE with test case checked. Uncle Owen is added at
that coordinate to monitor the interaction happened on that coordinate
and a tuskenRaider is added to attack Ben and see how he restore his
health.