Mind Control- Done Maybe?
- Luke is now able to mind control weak minded assuming that his force
ability is >=8. But the problem now is, there seems to be some delay in
the printing of the action.
  --> for example, when luke contols UncleOwen by asking UncleOwen to
take the canteen that exist on that coordinate, UncleOwen did already
took the canteen, but after the command return to Luke, the canteen
still exist there. But when luke attempts to pick up himself, there is a
null error, which means the canteen is not EVEN THERE! UncleOwen did
whatever action that Luke prompt.